[RCO-XXX] Allow the fix to happen for other projects, just the deletes are optional now.

+++ src/Zicht/Bundle/VersioningBundle/Command/ActivatePostponedVersionsCommand.php
@@ -113,7 +113,7 @@ class ActivatePostponedVersionsCommand extends ContainerAwareCommand
                 /** @var VersionableInterface $entity */
                 $entity = $this->doctrine->getManager()->find($scheduledVersion['source_class'], $scheduledVersion['original_id']);
 
-                if (!$entity && $isDeleteGhostEntries) {
+                if (!$entity) {
                     $this->getContainer()->get('logger')->warning(
                         sprintf(
                             'Entity %s with id %d does not exist anymore, purging record %s.',
@@ -132,8 +132,10 @@ class ActivatePostponedVersionsCommand extends ContainerAwareCommand
                         ));
                     }
 
-                    $stmt = $connection->prepare('DELETE FROM _entity_version WHERE id = :id');
-                    $stmt->execute([':id' => $scheduledVersion['id']]);
+                    if ($isDeleteGhostEntries) {
+                        $stmt = $connection->prepare('DELETE FROM _entity_version WHERE id = :id');
+                        $stmt->execute([':id' => $scheduledVersion['id']]);
+                    }
                     continue;
                 }
